+++ CHANGELOG.md
@@ -21,6 +21,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Add new directive interface `FieldBuilderDirective` https://github.com/nuwave/lighthouse/pull/1636
 - Add `@whereAuth` directive for filtering a field based on authenticated user https://github.com/nuwave/lighthouse/pull/1636
 - Use the `@trim` directive on fields to sanitize all input strings https://github.com/nuwave/lighthouse/pull/1641
+- Add Laravel Echo compatible subscription broadcaster https://github.com/nuwave/lighthouse/pull/1370
 - Allow auxiliary types in directive definitions https://github.com/nuwave/lighthouse/pull/1649
 
 ### Changed